Add a simple hello-world
Even reading along with *Rust for Rubyists* I still made a typo
resulting in the following compiler error:

  ❯ rustc rust_for_rubyists/hello.rs
  rust_for_rubyists/hello.rs:3:1: 3:3 error: expected `;` but found `fn`
  rust_for_rubyists/hello.rs:3 fn main() {
                               ^~

It took me longer than I'd like to admit to figure out what was going
on. It didn't occur to me that the `use` statement needed a semicolon —
the compiler error now makes total sense.